+++ client/components/order/OrderHistoryContainer.js
@@ -11,7 +11,7 @@ class OrderHistoryContainer extends React.Component {
 	}
 
 	render() {
-		let list = this.props.orders.map(order => <li><Link to={`/api/orders/${order.id}`} key={order.id}>{order.id}</Link></li>)
+		let list = this.props.orders.map(order => <li><Link to={`/orders/${order.id}`} key={order.id}>{order.id}</Link></li>)
 		return (
 			<div className="shoppingCartContainer marginTop">
 				<h2>OrderContainer</h2>
