+++ client/store/order.js
@@ -4,13 +4,11 @@ import axios from 'axios'
  * ACTION TYPES
  */
 const GOT_ORDERS = 'GOT_ORDERS'
-const GOT_ACTIVE_ORDER = 'GOT_ACTIVE_ORDER'
 
 /**
  * INITIAL STATE
  */
 const defaultOrders = []
-const activeOrder = {}
 
 /**
  * ACTION CREATORS
@@ -18,11 +16,6 @@ const activeOrder = {}
 const gotOrders = (orders) => ({
 	type: GOT_ORDERS,
 	orders
-})
-
-const gotActiveOrder = (activeOrder) => ({
-	type: GOT_ACTIVE_ORDER,
-	activeOrder
 })
 
 /**
@@ -37,15 +30,6 @@ export const fetchOrders = (userId) =>
 	})
 	.catch(err => console.log(err))
 
-export const fetchActiveOrder = (orderId) =>
-  dispatch =>
-	axios.get(`/api/orders/${orderId}/lineItems`)
-	.then(res => res.data)
-	.then(results => {
-		console.log(results)
-		dispatch(gotActiveOrder(results))
-	})
-
 
 /**
  * REDUCER
@@ -53,7 +37,7 @@ export const fetchActiveOrder = (orderId) =>
 export default function (state = defaultOrders, action) {
   switch (action.type) {
     case GOT_ORDERS:
-      return action.orders
+			return action.orders
     default:
       return state
   }
